pyproject.toml
[project]
name = "armature"
version = "0.1.0"
description = "Harness engineering framework for AI coding agents -- the invisible skeleton that shapes agent output"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["ai", "coding-agent", "harness", "linting", "quality", "architecture"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
python = ["ruff>=0.3.0", "mypy>=1.8.0", "pytest>=8.0.0"]
typescript = []
all = ["armature[python]", "armature[typescript]"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
armature = "armature.cli.main:cli"

[project.urls]
Homepage = "https://github.com/armature-dev/armature"
Repository = "https://github.com/armature-dev/armature"
Documentation = "https://github.com/armature-dev/armature#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/armature"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
